@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net7.0</TargetFramework>
 		<LangVersion>latest</LangVersion>
 		<GeneratePackageOnBuild>True</GeneratePackageOnBuild>
 		<Authors>Capgemini, ADCenter Valencia, Traiectum Team</Authors>
@@ -37,7 +37,7 @@
 	</ItemGroup>	
 	
 	<ItemGroup>
-    <PackageReference Include="AWSSDK.SQS" Version="3.7.2.74" />
+    <PackageReference Include="AWSSDK.SQS" Version="3.7.100.14" />
   </ItemGroup>
 
   <ItemGroup>
