@@ -151,7 +151,7 @@ namespace Devon4Net.Infrastructure.AWS.CDK.Resources.Handlers.Database
             });
         }
 
-        public IDatabaseInstance CreateDatabaseSqlServer(DeputyBase databaseEngineVersion, string identification, string instanceIdentifier, string userName, string password, StorageType storageType, InstanceClass instanceClass, string instanceSize, IVpc vpc, ISecurityGroup security, string securityGroupId, ISubnetGroup subnetGroup, string parameterGroupId = null, IRole[] parameterGroupRoles = null, double? allocatedStorageGb = 5, RemovalPolicy removalPolicy = RemovalPolicy.DESTROY, bool deleteAutomatedBackups = false, int backupRetentionDays = 1, bool? deletionProtection = false, SubnetType subnetType = SubnetType.PRIVATE_WITH_NAT, string defaultSubnetDomainSeparator = ",", string subnets = "", bool multiAZEnabled = true, bool? autoMinorVersionUpgrade = false, bool? storageEncrypted = true, string licenseOption = "LICENSE_INCLUDED", string edition = "ex", List<ISecurityGroup> securityGroups = null, IRole monitoringRole = null, int monitoringInterval = 0, bool enablePerformanceInsights = false, int? performanceInsightsRetention = 0) //NOSONAR number of params
+        public IDatabaseInstance CreateDatabaseSqlServer(DeputyBase databaseEngineVersion, string identification, string instanceIdentifier, string userName, string password, StorageType storageType, InstanceClass instanceClass, string instanceSize, IVpc vpc, ISecurityGroup security, string securityGroupId, ISubnetGroup subnetGroup, string parameterGroupId = null, IRole[] parameterGroupRoles = null, double? allocatedStorageGb = 5, RemovalPolicy removalPolicy = RemovalPolicy.DESTROY, bool deleteAutomatedBackups = false, int backupRetentionDays = 1, bool? deletionProtection = false, SubnetType subnetType = SubnetType.PRIVATE_WITH_EGRESS, string defaultSubnetDomainSeparator = ",", string subnets = "", bool multiAZEnabled = true, bool? autoMinorVersionUpgrade = false, bool? storageEncrypted = true, string licenseOption = "LICENSE_INCLUDED", string edition = "ex", List<ISecurityGroup> securityGroups = null, IRole monitoringRole = null, int monitoringInterval = 0, bool enablePerformanceInsights = false, int? performanceInsightsRetention = 0) //NOSONAR number of params
         {
             BasicDatabaseInfraWithHardcodedPassword(vpc, subnetType, defaultSubnetDomainSeparator, subnets, out var subnetSelection);
             var engine = GetInstanceEngine(databaseEngineVersion, edition);
