@@ -13,7 +13,7 @@ namespace Devon4Net.Infrastructure.AWS.CDK.Resources.Handlers.Database
         IDatabaseInstance CreateDatabase(DeputyBase databaseEngineVersion, string identification, string databaseName, double? port, string userName, ISecret passwordSecret, StorageType storageType, InstanceClass instanceClass, InstanceSize instanceSize, IVpc vpc, ISecurityGroup securityGroup, ISubnetGroup subnetGroup, IParameterGroup parameterGroup = null, double? allocatedStorageGb = 5, RemovalPolicy removalPolicy = RemovalPolicy.DESTROY, bool deleteAutomatedBackups = false, int backupRetentionDays = 1, bool deletionProtection = false, string[] logTypes = null, bool? storageEncrypted = null, bool? enableIamAuthentication = false, Duration enhancedMonitoringInterval = null, bool multiAZEnabled = true, bool autoMinorVersionUpgrade = false);//NOSONAR number of params
         IDatabaseInstance CreateDatabase(DeputyBase databaseEngineVersion, string identification, string databaseName, double? port, string userName, string secretName, StorageType storageType, InstanceClass instanceClass, InstanceSize instanceSize, IVpc vpc, string securityId, string securityGroupId, string parameterGroupId = null, IRole[] roles = null, double? allocatedStorageGb = 5, RemovalPolicy removalPolicy = RemovalPolicy.DESTROY, bool deleteAutomatedBackups = false, int backupRetentionDays = 1, bool deletionProtection = false, SubnetType subnetType = default, string defaultSubnetDomainSeparator = ",", string subnets = "", bool multiAZEnabled = true, bool autoMinorVersionUpgrade = false, bool? storageEncrypted = true);//NOSONAR number of params
         IDatabaseCluster CreateDatabaseCluster(DeputyBase databaseEngineVersion, string identification, string clusterIdentifier, string instanceIdentifierBase, string databaseName, double? port, double? instances, string userName, string secretName, IVpc vpc, InstanceClass instanceClass, InstanceSize instanceSize, string securityId, string securityGroupId, string parameterGroupId = null, IRole[] roles = null, bool storageEncrypted = true, SubnetType subnetType = default, string defaultSubnetDomainSeparator = ",", string subnets = "", RemovalPolicy removalPolicy = RemovalPolicy.DESTROY, int backupRetentionDays = 1, bool deletionProtection = false);//NOSONAR number of params
-        IDatabaseInstance CreateDatabaseSqlServer(DeputyBase databaseEngineVersion, string identification, string instanceIdentifier, string userName, string password, StorageType storageType, InstanceClass instanceClass, string instanceSize, IVpc vpc, ISecurityGroup security, string securityGroupId, ISubnetGroup subnetGroup, string parameterGroupId = null, IRole[] parameterGroupRoles = null, double? allocatedStorageGb = 5, RemovalPolicy removalPolicy = RemovalPolicy.DESTROY, bool deleteAutomatedBackups = false, int backupRetentionDays = 1, bool? deletionProtection = false, SubnetType subnetType = SubnetType.PRIVATE_WITH_NAT, string defaultSubnetDomainSeparator = ",", string subnets = "", bool multiAZEnabled = true, bool? autoMinorVersionUpgrade = false, bool? storageEncrypted = true, string licenseOption = "LICENSE_INCLUDED", string edition = "ex", List<ISecurityGroup> securityGroups = null, IRole monitoringRole = null, int monitoringInterval = 0, bool enablePerformanceInsights = false, int? performanceInsightsRetention = 0); //NOSONAR number of params
+        IDatabaseInstance CreateDatabaseSqlServer(DeputyBase databaseEngineVersion, string identification, string instanceIdentifier, string userName, string password, StorageType storageType, InstanceClass instanceClass, string instanceSize, IVpc vpc, ISecurityGroup security, string securityGroupId, ISubnetGroup subnetGroup, string parameterGroupId = null, IRole[] parameterGroupRoles = null, double? allocatedStorageGb = 5, RemovalPolicy removalPolicy = RemovalPolicy.DESTROY, bool deleteAutomatedBackups = false, int backupRetentionDays = 1, bool? deletionProtection = false, SubnetType subnetType = SubnetType.PRIVATE_WITH_EGRESS, string defaultSubnetDomainSeparator = ",", string subnets = "", bool multiAZEnabled = true, bool? autoMinorVersionUpgrade = false, bool? storageEncrypted = true, string licenseOption = "LICENSE_INCLUDED", string edition = "ex", List<ISecurityGroup> securityGroups = null, IRole monitoringRole = null, int monitoringInterval = 0, bool enablePerformanceInsights = false, int? performanceInsightsRetention = 0); //NOSONAR number of params
         IParameterGroup CreateParameterGroup(DeputyBase databaseEngineVersion, string identification, string parameterGroupName, IDictionary<string, string> parameters = null);//NOSONAR number of params
         IParameterGroup LocateParameterGroupByName(string identification, string parameterGroupName);
         ISubnet LocateSubnetById(string identification, string subnetId);
