@@ -5,7 +5,8 @@ using Amazon.SQS.Model;
 using Devon4Net.Infrastructure.AWS.Common.Options;
 using Devon4Net.Infrastructure.AWS.SQS.Dto;
 using Devon4Net.Infrastructure.AWS.SQS.Helper;
-using Devon4Net.Infrastructure.Logger.Logging;
+using Devon4Net.Infrastructure.AWS.SQS.Interfaces;
+using Devon4Net.Infrastructure.Common;
 using System;
 using System.Collections.Generic;
 using System.Linq;
