@@ -1,4 +1,4 @@
-﻿using Devon4Net.Infrastructure.Logger.Logging;
+﻿using Devon4Net.Infrastructure.Common;
 using FluentValidation;
 using FluentValidation.Results;
 using System;
