@@ -2,7 +2,7 @@
 using Confluent.Kafka.Admin;
 using Devon4Net.Infrastructure.Kafka.Exceptions;
 using Devon4Net.Infrastructure.Kafka.Options;
-using Devon4Net.Infrastructure.Logger.Logging;
+using Devon4Net.Infrastructure.Common;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 
