@@ -1,4 +1,4 @@
-﻿using Devon4Net.Infrastructure.Common.Common.IO;
+﻿using Devon4Net.Infrastructure.Common.IO;
 using Devon4Net.Infrastructure.LiteDb.LiteDb;
 using Devon4Net.Infrastructure.RabbitMQ.Options;
 using LiteDB;
