@@ -1,16 +1,18 @@
 ﻿using System.Net.Security;
-using Devon4Net.Infrastructure.Common.Common.IO;
 using Devon4Net.Infrastructure.Common.Handlers;
-using Devon4Net.Infrastructure.Extensions.Helpers;
+using Devon4Net.Infrastructure.Common.Helpers;
+using Devon4Net.Infrastructure.Common.Helpers.Interfaces;
+using Devon4Net.Infrastructure.Common.IO;
 using Devon4Net.Infrastructure.LiteDb.LiteDb;
 using Devon4Net.Infrastructure.LiteDb.Repository;
-using Devon4Net.Infrastructure.Logger.Logging;
+using Devon4Net.Infrastructure.Common;
 using Devon4Net.Infrastructure.RabbitMQ.Data.Service;
 using Devon4Net.Infrastructure.RabbitMQ.Domain.Database;
 using Devon4Net.Infrastructure.RabbitMQ.Domain.Entities;
 using Devon4Net.Infrastructure.RabbitMQ.Domain.ServiceInterfaces;
 using Devon4Net.Infrastructure.RabbitMQ.Options;
 using EasyNetQ;
+using EasyNetQ.DI;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Serilog;
@@ -99,8 +101,8 @@ namespace Devon4Net.Application.WebAPI.Configuration
 
         private static HostConfiguration GetHostConfiguration(HostDefinition host)
         {
-            var hostConfiguration = new HostConfiguration { Host = host.Host };
-            var port = (ushort) (host.Port != null ? (ushort) host.Port.Value : 0);
+            var port = (ushort)(host.Port != null ? (ushort)host.Port.Value : 0);
+            var hostConfiguration = new HostConfiguration { Host = host.Host, Port = port};
 
             if (port > 0) hostConfiguration.Port = port;
             _ = Enum.TryParse(host.SslPolicyErrors, out SslPolicyErrors sslPolicyErrors);
