@@ -2,7 +2,7 @@
 using System.Linq.Expressions;
 using Devon4Net.Domain.UnitOfWork.Exceptions;
 using Devon4Net.Domain.UnitOfWork.Pagination;
-using Devon4Net.Infrastructure.Logger.Logging;
+using Devon4Net.Infrastructure.Common;
 using Microsoft.EntityFrameworkCore;
 
 namespace Devon4Net.Domain.UnitOfWork.Repository
