@@ -1,18 +1,19 @@
-using Devon4Net.Application.WebAPI.Configuration;
-using Devon4Net.Application.WebAPI.Configuration.Application;
 using Devon4Net.Infrastructure.AWS.Serverless;
-using Devon4Net.Infrastructure.Logger;
-using Devon4Net.Infrastructure.Middleware.Middleware;
+using Devon4Net.Infrastructure.Cors;
 using Devon4Net.Infrastructure.Swagger;
+using Devon4Net.Infrastructure.Logger;
+using Devon4Net.Infrastructure.Common.Application.ApplicationTypes.API;
+using Devon4Net.Infrastructure.Common.Application.Middleware;
 
 var builder = WebApplication.CreateBuilder(args);
 
-builder.WebHost.InitializeDevonFw(builder.Host);
+builder.WebHost.InitializeDevonfwApi(builder.Host);
 
 #region services
+builder.Services.SetupDevonfwAws(builder.Configuration, true);
 builder.Services.AddControllers();
 builder.Services.AddEndpointsApiExplorer();
-builder.Services.ConfigureDevonfwAWS(builder.Configuration, true);
+builder.Services.SetupDevonfwAws(builder.Configuration, true);
 builder.Services.SetupMiddleware(builder.Configuration);
 builder.Services.SetupLog(builder.Configuration);
 builder.Services.SetupSwagger(builder.Configuration);
