@@ -3,7 +3,7 @@ using Devon4Net.Application.Kafka.Consumer.Business.FileManagement.Dto;
 using Devon4Net.Application.Kafka.Consumer.Business.FileManagement.Services;
 using Devon4Net.Infrastructure.Kafka.Handlers.Consumer;
 using Devon4Net.Infrastructure.Kafka.Options;
-using Devon4Net.Infrastructure.Logger.Logging;
+using Devon4Net.Infrastructure.Common;
 
 namespace Devon4Net.Application.Kafka.Consumer.Business.KafkaManagement.Handlers
 {
