@@ -1,42 +1,36 @@
 using Devon4Net.Application.WebAPI.Configuration;
-using Devon4Net.Application.WebAPI.Configuration.Application;
-using Devon4Net.Application.WebAPI.Implementation.Configuration;
 using Devon4Net.Domain.UnitOfWork;
 using Devon4Net.Infrastructure.CircuitBreaker;
+using Devon4Net.Infrastructure.Cors;
 using Devon4Net.Infrastructure.Grpc;
 using Devon4Net.Infrastructure.Kafka;
-using Devon4Net.Infrastructure.Logger;
-using Devon4Net.Infrastructure.Middleware.Middleware;
-using Devon4Net.Infrastructure.Nexus;
 using Devon4Net.Infrastructure.Swagger;
+using Devon4Net.Infrastructure.Logger;
+using Devon4Net.Infrastructure.Common.Application.ApplicationTypes.API;
+using Devon4Net.Infrastructure.Common.Application.Middleware;
 
 var builder = WebApplication.CreateBuilder(args);
 
-// Add services to the container.
-
-builder.WebHost.InitializeDevonFw(builder.Host);
-
+builder.WebHost.InitializeDevonfwApi(builder.Host);
 builder.Services.AddControllers();
-// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
-builder.Services.AddEndpointsApiExplorer();
 
 #region devon services
-var devonfwOptions =  builder.Services.SetupDevonfw(builder.Configuration);
+var devonfwOptions = builder.Services.SetupDevonfw(builder.Configuration);
+builder.Services.SetupLog(builder.Configuration!);
 builder.Services.SetupMiddleware(builder.Configuration);
-builder.Services.SetupLog(builder.Configuration);
 builder.Services.SetupSwagger(builder.Configuration);
 builder.Services.SetupCircuitBreaker(builder.Configuration);
 builder.Services.SetupCors(builder.Configuration);
 builder.Services.SetupJwt(builder.Configuration);
-builder.Services.SetupUnitOfWork(typeof(DevonConfiguration));
+builder.Services.SetupUnitOfWork(typeof(Program));
 builder.Services.SetupLiteDb(builder.Configuration);
 builder.Services.SetupRabbitMq(builder.Configuration);
 builder.Services.SetupMediatR(builder.Configuration);
-builder.Services.SetupNexus(builder.Configuration);
 builder.Services.SetupKafka(builder.Configuration);
 builder.Services.SetupGrpc(builder.Configuration);
-builder.Services.SetupDevonDependencyInjection(builder.Configuration);
 #endregion
+
+builder.Services.SetupCustomDependencyInjection(builder.Configuration);
 
 var app = builder.Build();
 
@@ -44,6 +38,7 @@ var app = builder.Build();
 app.ConfigureSwaggerEndPoint();
 app.SetupMiddleware(builder.Services);
 app.SetupCors();
+
 if (devonfwOptions.ForceUseHttpsRedirection || (!devonfwOptions.UseIIS && devonfwOptions.Kestrel.UseHttps))
 {
     app.UseHttpsRedirection();
@@ -51,6 +46,7 @@ if (devonfwOptions.ForceUseHttpsRedirection || (!devonfwOptions.UseIIS && devonf
 #endregion
 
 app.UseStaticFiles();
+app.UseHttpsRedirection();
 app.UseAuthorization();
 app.MapControllers();
 
